# apps/backend/src/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LinkedIn Agentic AI API",
    "version": "1.0.0",
    "description": "Complete API for LinkedIn Agentic AI system including agents, workflows, semantic search, analytics, feed, digests, providers, and posts."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/agent/post/write": {
      "post": {
        "summary": "Generate LinkedIn post draft",
        "tags": ["Agents"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topic": { "type": "string" },
                  "tone": { "type": "string" },
                  "industry": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Generated post" } }
      }
    },
...
    "/workflow/logs": {
      "get": {
        "summary": "Fetch workflow logs",
        "tags": ["Workflows"],
        "parameters": [
          { "in": "query", "name": "limit", "schema": { "type": "integer" } },
          { "in": "query", "name": "offset", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Workflow logs returned" } }
      }
    },
    "/posts": {
      "get": {
        "summary": "List saved posts",
        "tags": ["Posts"],
        "parameters": [
          { "in": "query", "name": "limit", "schema": { "type": "integer" } },
          { "in": "query", "name": "offset", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Posts returned" } }
      }
    },
    "/posts/{id}": {
      "get": {
        "summary": "Get a specific LinkedIn post",
        "tags": ["Posts"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Post returned" } }
      }
    },
...
  }
}